+++ ytrue-service/src/main/java/com/ytrue/serviceimpl/system/SysUserServiceImpl.java
@@ -1,10 +1,18 @@
 package com.ytrue.serviceimpl.system;
 
-import cn.hutool.core.convert.Convert;
+import cn.hutool.core.collection.CollUtil;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
+import com.ytrue.bean.dataobject.system.SysUser;
+import com.ytrue.bean.dataobject.system.SysUserJob;
+import com.ytrue.bean.dataobject.system.SysUserRole;
+import com.ytrue.bean.req.system.SysUserAddReq;
+import com.ytrue.bean.req.system.SysUserUpdatePasswordReq;
+import com.ytrue.bean.req.system.SysUserUpdateReq;
 import com.ytrue.bean.resp.system.SysUserDetailResp;
-import  com.ytrue.infra.core.response.ResponseCodeEnum;
+import com.ytrue.bean.resp.system.SysUserListResp;
+import com.ytrue.infra.core.constant.StrPool;
+import com.ytrue.infra.core.response.ResponseCodeEnum;
 import com.ytrue.infra.core.response.ServerResponseCode;
 import com.ytrue.infra.core.util.AssertUtil;
 import com.ytrue.infra.core.util.BeanUtil;
@@ -12,17 +20,15 @@ import com.ytrue.infra.db.base.BaseServiceImpl;
 import com.ytrue.infra.db.dao.system.SysUserDao;
 import com.ytrue.infra.db.dao.system.SysUserJobDao;
 import com.ytrue.infra.db.dao.system.SysUserRoleDao;
-import com.ytrue.bean.dataobject.system.SysUser;
-import com.ytrue.bean.dataobject.system.SysUserJob;
-import com.ytrue.bean.dataobject.system.SysUserRole;
-import com.ytrue.bean.req.system.SysUserReq;
-import com.ytrue.bean.resp.system.SysUserListResp;
-import com.ytrue.service.system.SysUserService;
 import com.ytrue.manager.DataScopeManager;
+import com.ytrue.service.system.SysUserService;
 import com.ytrue.tools.query.entity.QueryEntity;
 import com.ytrue.tools.query.enums.QueryMethod;
+import com.ytrue.tools.security.service.LoginService;
 import com.ytrue.tools.security.util.SecurityUtils;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.BeanUtils;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -42,18 +48,17 @@ public class SysUserServiceImpl extends BaseServiceImpl<SysUserDao, SysUser> imp
 
 
     private final SysUserRoleDao sysUserRoleDao;
-
     private final SysUserJobDao sysUserJobDao;
-
-
+    private final PasswordEncoder passwordEncoder;
     private final DataScopeManager dataScopeManager;
+    private final LoginService loginService;
 
     @Override
     public IPage<SysUserListResp> paginate(IPage<SysUserListResp> page, QueryEntity query) {
         // 处理数据过滤
         Set<Long> deptIds = dataScopeManager.listDeptIdDataScope();
         if (!deptIds.contains(0L)) {
-            query.addFilter(SysUser::getDeptId, QueryMethod.in, Convert.toList(deptIds), "u");
+            query.addFilter(SysUser::getDeptId, QueryMethod.in, deptIds.stream().toList(), "u");
         } else {
             query.addFilter(SysUser::getId, QueryMethod.eq, SecurityUtils.getLoginUser().getUser().getUserId(), "u");
         }
@@ -76,28 +81,35 @@ public class SysUserServiceImpl extends BaseServiceImpl<SysUserDao, SysUser> imp
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public void addUser(SysUserReq sysUserReq) {
-        SysUser user = BeanUtil.cgLibCopyBean(sysUserReq, SysUser::new);
+    public void addSysUser(SysUserAddReq requestParam) {
+        SysUser sysUser = new SysUser();
+        BeanUtils.copyProperties(requestParam, sysUser);
+
+        sysUser.setPassword(passwordEncoder.encode(StrPool.DEFAULT_PASSWORD));
+
         // 校验账号是否存在
-        AssertUtil.isNull(lambdaQuery().eq(SysUser::getUsername, sysUserReq.getUsername()).one(), ServerResponseCode.error("账号已存在"));
+        AssertUtil.isNull(lambdaQuery().eq(SysUser::getUsername, requestParam.getUsername()).one(), ServerResponseCode.error("账号已存在"));
         // 保存用户
-        save(user);
+        save(sysUser);
         // 保存用户与部门,角色的关系
-        sysUserReq.setId(user.getId());
-        saveRoleAndJobRelation(sysUserReq);
+        saveRoleAndJobRelation(sysUser.getId(), requestParam.getJobIds(), requestParam.getRoleIds());
     }
 
 
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public void updateUser(SysUserReq sysUserReq) {
-        SysUser user = BeanUtil.cgLibCopyBean(sysUserReq, SysUser::new);
+    public void updateSysUser(SysUserUpdateReq requestParam) {
+        SysUser sysUser = this.getById(requestParam.getId());
+        AssertUtil.notNull(sysUser, ResponseCodeEnum.ILLEGAL_OPERATION);
+
+        BeanUtils.copyProperties(requestParam, sysUser);
+
         // 删除用户与部门,角色的关系
-        deleteRoleAndJobRelation(Collections.singletonList(sysUserReq.getId()));
+        removeRoleAndJobRelation(Collections.singletonList(sysUser.getId()));
         // 更新角色
-        updateById(user);
+        updateById(sysUser);
         // 保存用户与菜单,部门的关系
-        saveRoleAndJobRelation(sysUserReq);
+        saveRoleAndJobRelation(sysUser.getId(), requestParam.getJobIds(), requestParam.getRoleIds());
     }
 
 
@@ -107,22 +119,45 @@ public class SysUserServiceImpl extends BaseServiceImpl<SysUserDao, SysUser> imp
         // 删除用户
         removeBatchByIds(ids);
         // 删除用户与部门,角色的关系
-        deleteRoleAndJobRelation(ids);
+        removeRoleAndJobRelation(ids);
+    }
+
+    @Override
+    public void updateSysUserPassword(SysUserUpdatePasswordReq requestParam) {
+        String userId = SecurityUtils.getLoginUser().getUser().getUserId();
+
+        SysUser sysUser = this.getById(Long.valueOf(userId));
+
+        AssertUtil.isTrue(passwordEncoder.matches(requestParam.getOldPassword(), sysUser.getPassword()), ServerResponseCode.error("旧密码错误"));
+        AssertUtil.objectNotEquals(requestParam.getOldPassword(), requestParam.getNewPassword(), ServerResponseCode.error("新密码不能与旧密码相同"));
+
+        sysUser.setPassword(passwordEncoder.encode(requestParam.getNewPassword()));
+        this.updateById(sysUser);
+        // 修改完成密码清除登录
+        loginService.logout(userId);
+    }
+
+    @Override
+    public void resetSysUserPassword(Long userId) {
+        this.lambdaUpdate().eq(SysUser::getId, userId).set(SysUser::getPassword, passwordEncoder.encode(StrPool.DEFAULT_PASSWORD)).update();
     }
 
     /**
      * 保存用户与部门,角色的关系
      *
-     * @param sysUserReq
+     * @param userIds
+     * @param jobIds
+     * @param roleIds
      */
-    private void saveRoleAndJobRelation(SysUserReq sysUserReq) {
-        //保存用户与岗位关系
-        if (!sysUserReq.getJobIds().isEmpty()) {
-            sysUserJobDao.insertBatchUserJob(sysUserReq.getId(), sysUserReq.getJobIds());
+    private void saveRoleAndJobRelation(Long userIds, Set<Long> jobIds, Set<Long> roleIds) {
+        //保存角色与菜单关系
+        if (CollUtil.isNotEmpty(jobIds)) {
+            sysUserJobDao.insertBatchUserJob(userIds, jobIds);
         }
-        //保存户与角色关系
-        if (!sysUserReq.getRoleIds().isEmpty()) {
-            sysUserRoleDao.insertBatchUserRole(sysUserReq.getId(), sysUserReq.getRoleIds());
+
+        //保存角色与部门关系
+        if (CollUtil.isNotEmpty(roleIds)) {
+            sysUserRoleDao.insertBatchUserRole(userIds, roleIds);
         }
     }
 
@@ -131,18 +166,13 @@ public class SysUserServiceImpl extends BaseServiceImpl<SysUserDao, SysUser> imp
      *
      * @param userIds
      */
-    private void deleteRoleAndJobRelation(List<Long> userIds) {
+    private void removeRoleAndJobRelation(List<Long> userIds) {
+        if (CollUtil.isEmpty(userIds)) {
+            return;
+        }
         //先删除用户与岗位关系
         sysUserJobDao.delete(Wrappers.<SysUserJob>lambdaQuery().in(SysUserJob::getUserId, userIds));
         //先删除用户与角色关系
         sysUserRoleDao.delete(Wrappers.<SysUserRole>lambdaQuery().in(SysUserRole::getUserId, userIds));
     }
-
-    @Override
-    public SysUser getUserByUsername(String username) {
-        SysUser user = lambdaQuery().eq(SysUser::getUsername, username).one();
-        AssertUtil.notNull(user, ResponseCodeEnum.DATA_NOT_FOUND);
-        return user;
-    }
-
 }
